+++ eroom-zoom-meetings-webinar/nuxy/helpers/helpers.php
@@ -32,6 +32,8 @@ function stm_wpcfto_nonces() {
 		'get_image_url',
 		'wpcfto_upload_file',
 		'wpcfto_search_posts',
+		'wpcfto_regenerate_fonts',
+		'wpcfto_create_term',
 	);
 
 	$nonces_list = array();