+++ eroom-zoom-meetings-webinar/nuxy/settings/settings.php
@@ -16,6 +16,10 @@ class WPCFTO_Settings {
 
 		add_action( 'admin_menu', array( $this, 'settings_page' ), 1000 );
 		add_action( 'wp_ajax_wpcfto_save_settings', array( $this, 'stm_save_settings' ) );
+		add_action( 'wp_ajax_wpcfto_create_term', array( $this, 'stm_create_term' ) );
+		add_action( 'wp_ajax_wpcfto_regenerate_fonts', array( $this, 'stm_regenerate_fonts' ) );
+		add_filter( 'wpcfto_enable_regenerate_fonts', array( $this, 'stm_enable_regenerate_fonts' ) );
+		add_filter( 'wpcfto_field_fonts_download_settings', array( $this, 'fonts_download_settings_template' ) );
 
 		if ( ! empty( $this->setup['admin_bar_title'] ) ) {
 			add_action( 'admin_bar_menu', array( $this, 'admin_bar_button' ), 40 );
@@ -30,12 +34,12 @@ class WPCFTO_Settings {
 		?>
 		<style>
 			<?php echo esc_attr( $selector ); ?>
-			img {
-				max-width: 25px;
-				vertical-align: top;
-				position: relative;
-				top: 3px;
-			}
+            img {
+                max-width: 25px;
+                vertical-align: top;
+                position: relative;
+                top: 3px;
+            }
 		</style>
 		<?php
 	}
@@ -60,6 +64,8 @@ class WPCFTO_Settings {
 	public function settings_page() {
 		if ( current_user_can( 'manage_options' ) ) {
 
+			$position = isset( $this->page_args['position'] ) ? $this->page_args['position'] : null;
+
 			if ( ! empty( $this->page_args['parent_slug'] ) ) {
 				$r = add_submenu_page(
 					$this->page_args['parent_slug'],
@@ -67,7 +73,8 @@ class WPCFTO_Settings {
 					$this->page_args['menu_title'],
 					'manage_options',
 					$this->page_args['menu_slug'],
-					array( $this, 'settings_page_view' )
+					array( $this, 'settings_page_view' ),
+					$position
 				);
 			} else {
 				add_menu_page(
@@ -77,7 +84,7 @@ class WPCFTO_Settings {
 					$this->page_args['menu_slug'],
 					array( $this, 'settings_page_view' ),
 					$this->page_args['icon'],
-					$this->page_args['position']
+					$position
 				);
 			}
 
@@ -127,13 +134,28 @@ class WPCFTO_Settings {
 	}
 
 	public function settings_page_view() {
-		$metabox               = $this->wpcfto_settings();
-		$settings              = $this->wpcfto_get_settings();
-		$page                  = $this->page_args;
-		$wpcfto_title          = ( ! empty( $this->setup['title'] ) ) ? $this->setup['title'] : '';
-		$wpcfto_sub_title      = ( ! empty( $this->setup['sub_title'] ) ) ? $this->setup['sub_title'] : '';
-		$wpcfto_logo           = ( ! empty( $this->setup['logo'] ) ) ? $this->setup['logo'] : STM_WPCFTO_URL . '/metaboxes/assets/images/stm-logo.svg';
-		$wpcfto_settings_alert = ( ! empty( $this->setup['save_settings_alert'] ) ) ? $this->setup['save_settings_alert'] : array(
+		$metabox                      = $this->wpcfto_settings();
+		$settings                     = $this->wpcfto_get_settings();
+		$page                         = $this->page_args;
+		$wpcfto_link                  = ( ! empty( $this->setup['additional_link'] ) && is_array( $this->setup['additional_link'] ) ) ? $this->setup['additional_link'] : array();
+		$link_text                    = ( ! empty( $wpcfto_link['text'] ) ) ? $wpcfto_link['text'] : '';
+		$link_icon                    = ( ! empty( $wpcfto_link['icon'] ) ) ? $wpcfto_link['icon'] : '';
+		$link_url                     = ( ! empty( $wpcfto_link['url'] ) ) ? $wpcfto_link['url'] : '';
+		$link_target                  = ( ! empty( $wpcfto_link['target'] ) ) ? $wpcfto_link['target'] : true;
+		$wpcfto_header_menu	          = ( ! empty( $this->setup['header_menu'] ) ) ? $this->setup['header_menu'] : array();
+		$wpcfto_header_menu_text      = ( ! empty( $wpcfto_header_menu['text'] ) ) ? $wpcfto_header_menu['text'] : '';
+		$wpcfto_header_menu_icon      = ( ! empty( $wpcfto_header_menu['icon'] ) ) ? $wpcfto_header_menu['icon'] : '';
+		$wpcfto_header_menu_url       = ( ! empty( $wpcfto_header_menu['url'] ) ) ? $wpcfto_header_menu['url'] : '';
+		$wpcfto_header_submenu        = ( ! empty( $this->setup['header_submenu'] ) ) ? $this->setup['header_submenu'] : array();
+		$wpcfto_header_submenu_text   = ( ! empty( $wpcfto_header_submenu['text'] ) ) ? $wpcfto_header_submenu['text'] : '';
+		$wpcfto_header_submenu_icon   = ( ! empty( $wpcfto_header_submenu['icon'] ) ) ? $wpcfto_header_submenu['icon'] : '';
+		$wpcfto_header_submenu_url    = ( ! empty( $wpcfto_header_submenu['url'] ) ) ? $wpcfto_header_submenu['url'] : '';
+		$wpcfto_header_submenu_target = ( ! empty( $wpcfto_header_submenu['target'] ) ) ? $wpcfto_header_submenu['target'] : true;
+
+		$wpcfto_title           = ( ! empty( $this->setup['title'] ) ) ? $this->setup['title'] : '';
+		$wpcfto_sub_title       = ( ! empty( $this->setup['sub_title'] ) ) ? $this->setup['sub_title'] : '';
+		$wpcfto_logo            = ( ! empty( $this->setup['logo'] ) ) ? $this->setup['logo'] : STM_WPCFTO_URL . '/metaboxes/assets/images/stm-logo.svg';
+		$wpcfto_settings_alert  = ( ! empty( $this->setup['save_settings_alert'] ) ) ? $this->setup['save_settings_alert'] : array(
 			'position'      => 'top_right',
 			'success_alert' => array(
 				'title'    => esc_html__( 'Saved!', 'nuxy' ),
@@ -173,11 +195,23 @@ class WPCFTO_Settings {
 		$id           = sanitize_text_field( $_REQUEST['name'] );
 		$settings     = array();
 		$request_body = file_get_contents( 'php://input' );
-		if ( ! empty( $request_body ) ) {
-			$request_body = json_decode( $request_body, true );
-			foreach ( $request_body as $section_name => $section ) {
+		if (!empty($request_body)) {
+			$request_body = json_decode($request_body, true);
+			$enable_download_font = $this->find_value_by_type($request_body, 'fonts_download_settings');
+			foreach ($request_body as $section_name => $section) {
 				foreach ( $section['fields'] as $field_name => $field ) {
-					$settings[ $field_name ] = $field['value'];
+					if ( $enable_download_font && class_exists( 'WPCFTO_WebFont_Loader' ) ) {
+						if ( ! empty( $field['value']['font-data']['family'] ) ) {
+							$exclude_font_family = ! empty( $field['excluded'] ) && in_array( 'font-family', $field['excluded'], true );
+							if ( ! $exclude_font_family ) {
+								$font                                     = new WPCFTO_WebFont_Loader( $field['value'], $field_name );
+								$field['value']['font-data']['local_url'] = $font->get_url();
+							}
+						}
+					}
+					if ( ! isset( $field['readonly'] ) || ! $field['readonly'] ) {
+						$settings[ $field_name ] = $field['value'];
+					}
 				}
 			}
 		}
@@ -196,6 +230,98 @@ class WPCFTO_Settings {
 		do_action( 'wpcfto_after_settings_saved', $id, $settings );
 
 		wp_send_json( $response );
+	}
+
+	public function stm_regenerate_fonts() {
+		check_ajax_referer( 'wpcfto_regenerate_fonts', 'nonce' );
+
+		if ( ! current_user_can( 'manage_options' ) && ! class_exists( 'WPCFTO_WebFont_Loader' ) ) {
+			die;
+		}
+
+		$id       = sanitize_text_field( $_GET['name'] );
+		$settings = get_option( $id, array() );
+
+		$response = array(
+			'reload'    => true,
+			'generated' => false,
+		);
+
+		$wpcfto_webfont = new WPCFTO_WebFont_Loader();
+
+		foreach ( $settings as $field_name => $field ) {
+			if ( ! empty( $field['font-data']['family'] ) && ! empty( $field['font-family'] ) ) {
+				$folder_name = $wpcfto_webfont->get_fonts_folder() . '/' . $field_name;
+				$wpcfto_webfont->deleteDirFiles( $folder_name );
+				$font                                              = new WPCFTO_WebFont_Loader( $field, $field_name );
+				$settings[ $field_name ]['font-data']['local_url'] = $font->get_url();
+			}
+		}
+		$response['generated'] = update_option( $this->option_name, $settings );
+
+		wp_send_json( $response );
+	}
+
+	public function stm_enable_regenerate_fonts( $val ) {
+		if ( ! current_user_can( 'manage_options' ) && ! class_exists( 'WPCFTO_WebFont_Loader' ) ) {
+			return false;
+		}
+
+		$settings = $this->wpcfto_get_settings();
+		if ( ! empty( $settings ) ) {
+			foreach ( $settings as $field_name => $field ) {
+				if ( ! empty( $field['font-data']['family'] ) ) {
+					return true;
+				}
+			}
+		}
+
+		return false;
+	}
+
+	public function stm_create_term() {
+		if ( ! current_user_can( 'manage_options' ) ) {
+			return false;
+		}
+
+		check_ajax_referer( 'wpcfto_create_term', 'nonce' );
+		$request_body = json_decode( file_get_contents( 'php://input' ) );
+
+		if ( empty( $request_body ) ) {
+			wp_send_json( array( 'error' => true, 'message' => 'Tag is empty' ) );
+		}
+
+		$taxonomy = sanitize_text_field( $request_body->new_taxonomy );
+		$term     = sanitize_text_field( $request_body->new_term );
+
+		if ( ! empty( $taxonomy ) && ! empty( $term ) ) {
+			$newTermData = wp_insert_term( $term, $taxonomy );
+
+			if ( $newTermData ) {
+				$term = get_term( $newTermData['term_id'], $taxonomy );
+				wp_send_json( array( 'success' => true, 'term' => $term ) );
+			}
+		}
+
+		wp_send_json( array( 'error' => true ) );
+	}
+
+	public function fonts_download_settings_template() {
+		return STM_WPCFTO_PATH . '/metaboxes/fields/fonts_download_settings.php';
+	}
+
+	public function find_value_by_type( $array, $type ) {
+		foreach ( $array as $tab ) {
+			if ( isset( $tab['fields'] ) ) {
+				foreach ( $tab['fields'] as $field ) {
+					if ( isset( $field['type'] ) && $field['type'] === $type ) {
+						return $field['value'];
+					}
+				}
+			}
+		}
+
+		return false;
 	}
 }